init software stack and refine structure
